Move instr to src
