--- src/new/mod.rs
@@ -1,2 +1,0 @@
-pub mod nodes;
-pub mod instr;
